Retrieve Docker tags for a Rocker image

Queries the Docker Hub API to retrieve all available tags for a specified Rocker image. Supports user-friendly modes: "base", "rstudio", and "tidyverse". Returns a structured list containing the image name, tag vector, and source URL.

Usage

get_r_ver_tags(r_mode = "base", verbose = FALSE)

Arguments

r_mode

Character string. One of "base", "rstudio", or "tidyverse". Determines which Rocker image to query. "base" maps to "rocker/r-ver".

verbose

Logical. If TRUE, prints progress messages during tag retrieval and pagination.

Value

A named list with the following elements:

image

Character string. The full Docker image name, e.g. "rocker/r-ver".

tags

Character vector. All available tags for the specified image, e.g. c("latest", "devel", "4.4", "4.4.3", ...).

source

Character string. The base URL of the Docker Hub API used to retrieve the tags.

Examples

get_r_ver_tags("base")
#> $image
#> [1] "rocker/r-ver"
#> 
#> $tags
#>  [1] "latest"                     "devel"                     
#>  [3] "4.4"                        "4.4.3"                     
#>  [5] "4"                          "4.5"                       
#>  [7] "4.5.1"                      "4.5.0"                     
#>  [9] "4.3"                        "4.3.3"                     
#> [11] "4.4.2"                      "4.4.1"                     
#> [13] "4.2"                        "4.2.3"                     
#> [15] "4.4.0"                      "4.3.0"                     
#> [17] "4.3.1"                      "4.3.2"                     
#> [19] "4.1"                        "4.1.3"                     
#> [21] "4.1.3-cuda11.1"             "4.0"                       
#> [23] "4.0.5"                      "4.2.1-cuda10.1"            
#> [25] "4.2.2"                      "4.2.1"                     
#> [27] "4.2.0"                      "4.2.0-cuda11.1"            
#> [29] "4.2.1-cuda11.1"             "4.1.3-cuda10.1"            
#> [31] "4.0.5-cuda11.1"             "4.0.5-cuda10.1"            
#> [33] "4.2.0-cuda10.1"             "4.1.0-cuda10.1"            
#> [35] "4.1.1-cuda10.1"             "4.1.2-cuda10.1"            
#> [37] "4.1.1"                      "4.1.2"                     
#> [39] "4.1.0"                      "4.0.4-cuda11.1"            
#> [41] "4.0.3-cuda11.1"             "4.0.2-cuda10.1"            
#> [43] "4.0.4-cuda10.1"             "4.0.3-cuda10.1"            
#> [45] "4.0.0-cuda10.1"             "4.0.1-cuda10.1"            
#> [47] "4.1.0-cuda11.1"             "4.0.2"                     
#> [49] "4.1.1-cuda11.1"             "4.1.2-cuda11.1"            
#> [51] "4.0.4"                      "4.0.1"                     
#> [53] "4.0.3"                      "4.0.0"                     
#> [55] "4.0.0-ubuntu18.04"          "devel-cuda11.1"            
#> [57] "devel-cuda11.3"             "devel-cuda10.1"            
#> [59] "3.6.3-ubuntu18.04"          "4.0.0-cuda10.2-ubuntu18.04"
#> [61] "4.0.0-cuda10.0-ubuntu18.04" "3.6.3"                     
#> [63] "3.6.2"                      "3.6.1"                     
#> [65] "3.5.1"                      "3.6.0"                     
#> [67] "3.5.3"                      "3.5.0"                     
#> [69] "3.4.4"                      "3.4.3"                     
#> [71] "3.5.2"                      "3.4.2"                     
#> [73] "3.4.1"                      "3.4.0"                     
#> [75] "3.3.3"                      "3.3.2"                     
#> [77] "3.2.5"                      "3.3.0"                     
#> [79] "3.3.1"                      "3.2.4"                     
#> [81] "3.2.3"                      "3.2.0"                     
#> [83] "3.2.2"                      "3.2.1"                     
#> [85] "3.1.3"                      "3.1.2"                     
#> [87] "3"                          "3.6"                       
#> [89] "3.1"                        "3.1.0"                     
#> [91] "3.3"                        "3.5"                       
#> [93] "3.4"                        "3.2"                       
#> [95] "devel-san"                 
#> 
#> $source
#> [1] "https://hub.docker.com/v2/repositories"
#> 
get_r_ver_tags("rstudio", verbose = TRUE)
#> Fetching tags from: https://hub.docker.com/v2/repositories/rocker/rstudio/tags?page_size=100
#> $image
#> [1] "rocker/rstudio"
#> 
#> $tags
#>  [1] "latest"            "devel"             "4.4"              
#>  [4] "4"                 "4.4.3"             "4.5"              
#>  [7] "4.5.1"             "4.5.0"             "4.3"              
#> [10] "4.3.3"             "4.4.2"             "4.4.1"            
#> [13] "4.4.0"             "4.3.0"             "4.3.1"            
#> [16] "4.3.2"             "4.2"               "4.2.3"            
#> [19] "4.1"               "4.1.3"             "4.0"              
#> [22] "4.0.5"             "4.2.2"             "4.2.1"            
#> [25] "4.2.0"             "latest-daily"      "4.1.1"            
#> [28] "4.1.2"             "4.1.0"             "4.0.2"            
#> [31] "4.0.4"             "4.0.1"             "4.0.3"            
#> [34] "4.0.0"             "4.0.0-ubuntu18.04" "4.0.5-daily"      
#> [37] "4.0.4-daily"       "4.0.3-daily"       "3.6.3-ubuntu18.04"
#> [40] "3.6.3"             "3.6.2"             "3.2.0"            
#> [43] "3.2.5"             "3.3.0"             "3.6.1"            
#> [46] "3"                 "3.6"               "3.6.0"            
#> [49] "3.5"               "3.5.3"             "3.5.2"            
#> [52] "3.5.1"             "3.4"               "3.4.4"            
#> [55] "3.3.1"             "3.4.0"             "3.5.0"            
#> [58] "3.4.1"             "3.4.3"             "3.3"              
#> [61] "3.3.3"             "3.4.2"             "3.3.2"            
#> [64] "testing"           "sandbox"          
#> 
#> $source
#> [1] "https://hub.docker.com/v2/repositories"
#>